+++ extension/src/python/index.ts
@@ -4,7 +4,6 @@ import { getProcessPlatform } from '../env'
 import { exists } from '../fileSystem'
 import { Logger } from '../common/logger'
 import { createProcess, executeProcess, Process } from '../process/execution'
-import { esmPackagesImported } from '../util/esm'
 
 const sendOutput = (process: Process) => {
   process.all?.on('data', chunk =>
@@ -35,7 +34,6 @@ export const setupTestVenv = async (
   envDir: string,
   ...installArgs: string[]
 ) => {
-  await esmPackagesImported
   if (!exists(join(cwd, envDir))) {
     const initVenv = createProcess({
       args: ['-m', 'venv', envDir],