+++ extension/src/setup/webview/messages.ts
@@ -9,7 +9,7 @@ import { BaseWebview } from '../../webview'
 import { sendTelemetryEvent } from '../../telemetry'
 import { EventName } from '../../telemetry/constants'
 import { selectPythonInterpreter } from '../../extensions/python'
-import { autoInstallDvc } from '../autoInstall'
+import { autoInstallDvc, autoUpgradeDvc } from '../autoInstall'
 import {
   RegisteredCliCommands,
   RegisteredCommands
@@ -42,13 +42,15 @@ export class WebviewMessages {
     pythonBinPath,
     remoteList,
     sectionCollapsed,
-    shareLiveToStudio
+    shareLiveToStudio,
+    isAboveLatestTestedVersion
   }: SetupData) {
     void this.getWebview()?.show({
       canGitInitialize,
       cliCompatible,
       dvcCliDetails,
       hasData,
+      isAboveLatestTestedVersion,
       isPythonExtensionUsed,
       isStudioConnected,
       needsGitCommit,
@@ -77,6 +79,8 @@ export class WebviewMessages {
         return this.selectPythonInterpreter()
       case MessageFromWebviewType.INSTALL_DVC:
         return this.installDvc()
+      case MessageFromWebviewType.UPGRADE_DVC:
+        return this.upgradeDvc()
       case MessageFromWebviewType.SETUP_WORKSPACE:
         return commands.executeCommand(
           RegisteredCommands.EXTENSION_SETUP_WORKSPACE
@@ -129,6 +133,12 @@ export class WebviewMessages {
       undefined
     )
     return selectPythonInterpreter()
+  }
+
+  private upgradeDvc() {
+    sendTelemetryEvent(EventName.VIEWS_SETUP_UPGRADE_DVC, undefined, undefined)
+
+    return autoUpgradeDvc()
   }
 
   private installDvc() {