+++ extension/src/test/suite/setup/index.test.ts
@@ -139,6 +139,22 @@ suite('Setup Test Suite', () => {
       expect(mockAutoInstallDvc).to.be.calledOnce
     }).timeout(WEBVIEW_TEST_TIMEOUT)
 
+    it('should handle an auto upgrade dvc message from the webview', async () => {
+      const { messageSpy, setup, mockAutoUpgradeDvc } = buildSetup(disposable)
+
+      const webview = await setup.showWebview()
+      await webview.isReady()
+
+      const mockMessageReceived = getMessageReceivedEmitter(webview)
+
+      messageSpy.resetHistory()
+      mockMessageReceived.fire({
+        type: MessageFromWebviewType.UPGRADE_DVC
+      })
+
+      expect(mockAutoUpgradeDvc).to.be.calledOnce
+    }).timeout(WEBVIEW_TEST_TIMEOUT)
+
     it('should handle a select Python interpreter message from the webview', async () => {
       const { messageSpy, mockExecuteCommand, setup } = buildSetup(disposable)
       const setInterpreterCommand = 'python.setInterpreter'
@@ -241,6 +257,7 @@ suite('Setup Test Suite', () => {
         cliCompatible: undefined,
         dvcCliDetails: { command: 'dvc', version: undefined },
         hasData: false,
+        isAboveLatestTestedVersion: undefined,
         isPythonExtensionUsed: false,
         isStudioConnected: false,
         needsGitCommit: true,
@@ -283,6 +300,7 @@ suite('Setup Test Suite', () => {
         cliCompatible: true,
         dvcCliDetails: { command: 'dvc', version: MIN_CLI_VERSION },
         hasData: false,
+        isAboveLatestTestedVersion: false,
         isPythonExtensionUsed: false,
         isStudioConnected: false,
         needsGitCommit: true,
@@ -334,6 +352,7 @@ suite('Setup Test Suite', () => {
           version: MIN_CLI_VERSION
         },
         hasData: false,
+        isAboveLatestTestedVersion: false,
         isPythonExtensionUsed: false,
         isStudioConnected: false,
         needsGitCommit: false,
@@ -385,6 +404,7 @@ suite('Setup Test Suite', () => {
           version: MIN_CLI_VERSION
         },
         hasData: false,
+        isAboveLatestTestedVersion: false,
         isPythonExtensionUsed: false,
         isStudioConnected: false,
         needsGitCommit: true,