+++ webview/src/setup/components/App.tsx
@@ -12,6 +12,7 @@ import { SetupContainer } from './SetupContainer'
 import { Remotes } from './remote/Remotes'
 import { useVsCodeMessaging } from '../../shared/hooks/useVsCodeMessaging'
 import { sendMessage } from '../../shared/vscode'
+import { TooltipIconType } from '../../shared/components/sectionContainer/InfoTooltip'
 import { SetupDispatch, SetupState } from '../store'
 import {
   updateSectionCollapsed,
@@ -21,6 +22,7 @@ import {
   updateCanGitInitialize,
   updateCliCompatible,
   updateDvcCliDetails,
+  updateIsAboveLatestTestedVersion,
   updateIsPythonExtensionUsed,
   updateNeedsGitInitialized,
   updateProjectInitialized,
@@ -36,6 +38,26 @@ import {
   updateShareLiveToStudio
 } from '../state/studioSlice'
 
+const getDvcStatusIcon = (
+  isDvcSetup: boolean,
+  isVersionAboveLatestTested: boolean
+) => {
+  if (!isDvcSetup) {
+    return TooltipIconType.ERROR
+  }
+
+  return isVersionAboveLatestTested
+    ? TooltipIconType.WARNING
+    : TooltipIconType.PASSED
+}
+
+const getStudioStatusIcon = (cliCompatible: boolean, isConnected: boolean) => {
+  if (!cliCompatible) {
+    return TooltipIconType.ERROR
+  }
+
+  return isConnected ? TooltipIconType.PASSED : TooltipIconType.INFO
+}
 export const feedStore = (
   data: MessageToWebview<SetupData>,
   dispatch: SetupDispatch
@@ -60,6 +82,11 @@ export const feedStore = (
         continue
       case 'isPythonExtensionUsed':
         dispatch(updateIsPythonExtensionUsed(data.data.isPythonExtensionUsed))
+        continue
+      case 'isAboveLatestTestedVersion':
+        dispatch(
+          updateIsAboveLatestTestedVersion(data.data.isAboveLatestTestedVersion)
+        )
         continue
       case 'isStudioConnected':
         dispatch(updateIsStudioConnected(data.data.isStudioConnected))
@@ -93,9 +120,8 @@ export const feedStore = (
 }
 
 export const App: React.FC = () => {
-  const { projectInitialized, cliCompatible } = useSelector(
-    (state: SetupState) => state.dvc
-  )
+  const { projectInitialized, cliCompatible, isAboveLatestTestedVersion } =
+    useSelector((state: SetupState) => state.dvc)
   const hasExperimentsData = useSelector(
     (state: SetupState) => state.experiments.hasData
   )
@@ -131,29 +157,46 @@ export const App: React.FC = () => {
       <SetupContainer
         sectionKey={SetupSection.DVC}
         title="DVC"
-        isSetup={isDvcSetup}
+        icon={getDvcStatusIcon(isDvcSetup, !!isAboveLatestTestedVersion)}
+        overrideSectionDescription={
+          isAboveLatestTestedVersion ? (
+            <>
+              The located version has not been tested against the extension. If
+              you are experiencing unexpected behaviour, first try to update the
+              extension. If there are no updates available, please downgrade DVC
+              to the same minor version as displayed and reload the window.
+            </>
+          ) : undefined
+        }
       >
         <Dvc />
       </SetupContainer>
       <SetupContainer
         sectionKey={SetupSection.EXPERIMENTS}
         title="Experiments"
-        isSetup={isDvcSetup && !!hasExperimentsData}
+        icon={
+          isDvcSetup && hasExperimentsData
+            ? TooltipIconType.PASSED
+            : TooltipIconType.ERROR
+        }
       >
         <Experiments isDvcSetup={projectInitialized && !!cliCompatible} />
       </SetupContainer>
       <SetupContainer
         sectionKey={SetupSection.REMOTES}
         title="Remotes"
-        isSetup={!!(remoteList && Object.values(remoteList).some(Boolean))}
+        icon={
+          remoteList && Object.values(remoteList).some(Boolean)
+            ? TooltipIconType.PASSED
+            : TooltipIconType.ERROR
+        }
       >
         <Remotes cliCompatible={!!cliCompatible} remoteList={remoteList} />
       </SetupContainer>
       <SetupContainer
         sectionKey={SetupSection.STUDIO}
         title="Studio"
-        isSetup={!!cliCompatible}
-        isConnected={isStudioConnected}
+        icon={getStudioStatusIcon(!!cliCompatible, isStudioConnected)}
       >
         <Studio
           setShareLiveToStudio={setShareLiveToStudio}