+++ webview/src/setup/components/remote/ProjectRemotes.tsx
@@ -5,11 +5,13 @@ export const ProjectRemotes: React.FC<{
   remotes: { [alias: string]: string }
 }> = ({ remotes }) => (
   <table data-testid="remote-details" className={styles.remoteDetails}>
-    {Object.entries(remotes).map(([alias, remote]) => (
-      <tr key={[alias, remote].join('-')}>
-        <th className={styles.alias}>{alias}</th>
-        <th className={styles.remote}>{remote}</th>
-      </tr>
-    ))}
+    <tbody>
+      {Object.entries(remotes).map(([alias, remote]) => (
+        <tr key={[alias, remote].join('-')}>
+          <th className={styles.alias}>{alias}</th>
+          <th className={styles.remote}>{remote}</th>
+        </tr>
+      ))}
+    </tbody>
   </table>
 )